Fix responsive layout of homepage
Replaced absolute units with relative units. Replaced unordered list for
business hours with tables to make it more uniform in appearance.
Removed opacity and replaced the background color with an opaque white
so that the content boxes are opaque without affect the text color or
any other content within the content boxes.