Update responsive layout and CSS styling
Updated the font size, margin, and padding of elements on menu-related
web pages. Created span elements and applied flex display to improve
layout display of menu items. Used dotted bottom border of an empty span
element to create the effect of ellipses for menu items to display
prices in a more uniform and easily discernable format.