Create modal for alerts and update menu content
Updated the lunch and dinner appetizer web pages with new information
regarding a menu item. Created a modal or popup box advertising the menu
change. Styled the modal with CSS, programmed the box to appear on
on screen one second after homepage loads so that users realize it is a
popup box. Applied an event listener to the "close" div element to allow
users to close the popup box to better navigate the page after reading
the popup's content. Updated the business hours section to state a
special close-for-business date.