Add modal to menu web page and update index style
Created a new popup window to bring attention to a change in schedule on
the menu web page in case users ignore the business hours section on the
homepage. Changed the font-size for business days in the business hours
section to keep hours on one line on smaller screens. Updated the home-
page modal style to ensure information is more visible.